modificações de inimigo movel

+++ Cobra.py
@@ -10,7 +10,7 @@ class Cobra(object):
         # Tamanho da cobra
         self.tamanho = 1
         # Posição de toda a cobra, começando no meio da tela
-        self.posicoes = [((v.largura / 2), (v.altura / 2))]
+        self.posicoes = [(int(v.largura / 2), int(v.altura / 2))]
         # Começar se movendo numa direção aleatória
         self.direcao = random.choice([v.CIMA, v.BAIXO, v.ESQUERDA, v.DIREITA])
         # Cor da cobra
@@ -37,7 +37,7 @@ class Cobra(object):
         x, y = self.direcao
         # Próxima localização da cobra
         # Qualquer que seja a direção, irá se mover apenas um quadrado por vez
-        novo = (((atual[0] + (x * v.tamanho_rede)) % v.largura), (atual[1] + (y * v.tamanho_rede)) % v.altura)
+        novo = (int((atual[0] + (x * v.tamanho_rede)) % v.largura), int(atual[1] + (y * v.tamanho_rede)) % v.altura)
         # Fazer com que o resto do corpo se mova para o lugar em que o pedaço da frente estava
         # Movendo apenas a "cabeça" e a "cauda"
         if len(self.posicoes) > 2 and novo in self.posicoes[2:]:
